+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,12 +1,15 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.TableGroupService;
-import kitchenpos.application.dto.TableGroupCreateRequest;
-import kitchenpos.domain.TableGroup;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
 import java.net.URI;
+import kitchenpos.application.TableGroupService;
+import kitchenpos.domain.TableGroup;
+import kitchenpos.ui.dto.TableGroupCreateRequest;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class TableGroupRestController {
